now support float16

+++ lib/model_zoo/openaimodel.py
@@ -2527,7 +2527,7 @@ class UNetModelVD(nn.Module):
     def forward_dc(self, x, timesteps, c0, c1, xtype, c0_type, c1_type, mixed_ratio):
         hs = []
         t_emb = timestep_embedding(timesteps, self.model_channels, repeat_only=False)
-        emb = self.time_embed(t_emb)
+        emb = self.time_embed(t_emb.to(x.dtype))
 
         if xtype == 'text':
             x = x[:, :, None, None]
